requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cachetools==4.2.2
certifi==2020.12.5
cfgv==3.3.0
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.3
django-heroku==0.3.1
filelock==3.0.12
google-api-core==1.28.0
google-auth==1.30.0
google-cloud-dialogflow==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
identify==2.2.5
idna==2.10
nodeenv==1.6.0
packaging==20.9
pre-commit==2.13.0
proto-plus==1.18.1
protobuf==3.17.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.6
whitenoise==5.2.0
